test(goxi): split the birth-quota decision tests into the admission module
`process_lease_tests.rs` crossed the source-size guard's BYTE ceiling
(52332 > MAX_BYTES=51200) while still under MAX_LINES, which is how this guard
usually trips: the line count looks fine.

Move the four decision -> birth-quota tests — `shadow_epoch_binds_but_never_lifts`,
`unleased_epoch_is_born_unclamped_because_no_grant_can_ever_lift_it`,
`birth_authority_is_armed_only_when_a_lift_is_reachable`, and
`absent_handoff_row_is_unleased_and_therefore_unarmed` — into
`process_lease_admission_tests.rs`, which is where they belong: every one of them
is about which admission decision reaches the launcher and what quota it commits.

No assertion is deleted, weakened, or ignored, and no size-guard threshold or
exemption marker is touched. Both files are now under both limits
(45743 and 16215 bytes), and `lease_runner_tests` still enumerates the same 41
tests it did before the move.

Doc cross-references in `evaluate_invocation_lift` updated to the new path.
