+++ server/crates/djinn-agent/src/process/tests/process_lease_admission_tests.rs
@@ -210,3 +210,169 @@ async fn a_non_platform_database_fails_closed_instead_of_lifting() {
         "a failed epoch read must never lift cpu.max"
     );
 }
+
+// ---------------------------------------------------------------------------
+// Decision -> birth-quota mapping, and the two decisions that must NOT lift.
+//
+// Moved here from `process_lease_tests.rs` (which was over the 51200-byte source
+// guard) because they belong with the tests above: every one of them is about
+// which admission decision reaches the launcher and what quota it commits.
+// ---------------------------------------------------------------------------
+
+/// AC1: a shadow epoch never lifts cpu.max even on a valid matching bind; the
+/// spawn still traverses the launcher and the durable lease is reconciled.
+#[tokio::test]
+async fn shadow_epoch_binds_but_never_lifts() {
+    let services = Arc::new(ScriptedServices::new(
+        vec![granted(7)],
+        vec![status(LeaseState::Active, Some(7))],
+        vec![status(LeaseState::Active, Some(7)); 20],
+    ));
+    services.set_lift_decision(djinn_supervisor::services::InvocationLiftDecision::Shadow);
+    services
+        .release
+        .lock()
+        .unwrap()
+        .push_back(LeaseResult::Released {
+            candidate_cleanup: false,
+        });
+    let launcher = Arc::new(ScriptedLauncher::default());
+    let cancel = CancellationToken::new();
+    let runner = LeaseInvocationRunner::new(
+        services.clone(),
+        services.clone(),
+        launcher.clone(),
+        clock(),
+    );
+    let run_cancel = cancel.clone();
+    let run = tokio::spawn(async move { runner.output(command(), config(), run_cancel).await });
+    wait_for(&services.status_calls, 3).await;
+    cancel.cancel();
+    run.await.unwrap().unwrap();
+    // The launcher was driven (queue + grant happened) but never lifted.
+    assert_eq!(services.queue_calls.load(Ordering::SeqCst), 1);
+    assert_eq!(services.grant_calls.load(Ordering::SeqCst), 1);
+    assert!(
+        launcher.lifts.lock().unwrap().is_empty(),
+        "shadow epoch must never lift cpu.max"
+    );
+    // Shadow is the one decision that clamps ON PURPOSE: it is an observation
+    // mode, so the leaf must still be born at the unleased quota.
+    assert_eq!(
+        launcher.authorities(),
+        vec![djinn_cgroup_launcher::LeaseAuthority::Armed],
+        "shadow observation is only meaningful against a clamped leaf"
+    );
+    // The durable lease is still reconciled to terminal (fence recorded).
+    assert!(services.release_calls.load(Ordering::SeqCst) <= 1);
+    assert_eq!(launcher.kills.load(Ordering::SeqCst), 1);
+}
+
+/// AC3 (agent side): an unknown/stale/baseline/ABSENT epoch cannot lift, so the
+/// leaf must be born WITHOUT a quota of its own — not pinned to the unleased
+/// quota for the whole command.
+///
+/// # The production symptom this now names (goxi launcher blocker 11)
+///
+/// The previous version of this test asserted only `lifts.is_empty()` and
+/// passed, because a no-op lift is exactly what the code did. What it could not
+/// see is that the leaf had *already* been born at 250m, so "never lifts" meant
+/// "clamped at 250m forever". Production ran with the durable
+/// `admission_handoff` row absent — `Unleased` for every invocation — and a
+/// measured leaf reached 21.1 CPU-seconds, 84x the 0.25 CPU-s escalation
+/// threshold, with `cpu.max` still reading `25000 100000`. Builds ran ~16x
+/// slower armed than disabled.
+///
+/// Reverting `birth_authority`'s `Unleased` arm to `Armed` reproduces that:
+/// this assertion fails with `Armed`, i.e. "the leaf was clamped by an authority
+/// that can never lift it".
+#[tokio::test]
+async fn unleased_epoch_is_born_unclamped_because_no_grant_can_ever_lift_it() {
+    let services = Arc::new(ScriptedServices::new(
+        vec![granted(7)],
+        vec![status(LeaseState::Active, Some(7))],
+        vec![status(LeaseState::Active, Some(7)); 20],
+    ));
+    services.set_lift_decision(djinn_supervisor::services::InvocationLiftDecision::Unleased);
+    services
+        .release
+        .lock()
+        .unwrap()
+        .push_back(LeaseResult::Released {
+            candidate_cleanup: false,
+        });
+    let launcher = Arc::new(ScriptedLauncher::default());
+    let cancel = CancellationToken::new();
+    let runner = LeaseInvocationRunner::new(
+        services.clone(),
+        services.clone(),
+        launcher.clone(),
+        clock(),
+    );
+    let run_cancel = cancel.clone();
+    let run = tokio::spawn(async move { runner.output(command(), config(), run_cancel).await });
+    wait_for(&services.status_calls, 3).await;
+    cancel.cancel();
+    run.await.unwrap().unwrap();
+    assert_eq!(services.grant_calls.load(Ordering::SeqCst), 1);
+    assert!(
+        launcher.lifts.lock().unwrap().is_empty(),
+        "unleased epoch must never lift cpu.max"
+    );
+    assert_eq!(
+        launcher.authorities(),
+        vec![djinn_cgroup_launcher::LeaseAuthority::Unarmed],
+        "an epoch that can never grant a lift must not clamp the leaf either; \
+         `Armed` here IS the production defect (cpu.max pinned at 25000 100000 \
+         while the leaf burned 21.1 CPU-seconds)"
+    );
+    assert!(services.release_calls.load(Ordering::SeqCst) <= 1);
+}
+
+/// The mapping itself, enumerated: every decision the durable authority can
+/// produce, and the birth quota it commits.
+///
+/// `Unleased` is what `evaluate_invocation_lift` returns for an ABSENT handoff
+/// row, which is the state production is actually in — see
+/// `absent_handoff_row_is_unleased_and_therefore_unarmed` below for the
+/// composition that ties the two together.
+#[test]
+fn birth_authority_is_armed_only_when_a_lift_is_reachable() {
+    use djinn_cgroup_launcher::LeaseAuthority;
+    use djinn_supervisor::services::InvocationLiftDecision;
+    assert_eq!(
+        crate::process::birth_authority(InvocationLiftDecision::Lift),
+        LeaseAuthority::Armed
+    );
+    assert_eq!(
+        crate::process::birth_authority(InvocationLiftDecision::Shadow),
+        LeaseAuthority::Armed
+    );
+    assert_eq!(
+        crate::process::birth_authority(InvocationLiftDecision::Unleased),
+        LeaseAuthority::Unarmed
+    );
+}
+
+/// The composition that failed in production: the DURABLE row state that
+/// production actually has, projected through the real `evaluate_invocation_lift`
+/// and the real `birth_authority`, must reach the launcher as `Unarmed`.
+///
+/// `djinn-server epoch show` on the production node reported
+/// `admission handoff row: <absent>` while the launcher was armed. That is not a
+/// hypothetical: it is the exact input below.
+#[test]
+fn absent_handoff_row_is_unleased_and_therefore_unarmed() {
+    use djinn_cgroup_launcher::LeaseAuthority;
+    use djinn_supervisor::services::evaluate_invocation_lift;
+    // Absent row (production), and unreadable row (a DB blip) — both fail closed
+    // on the lift, so neither may clamp.
+    for row in [Ok(None), Err(())] {
+        let decision = evaluate_invocation_lift(row);
+        assert_eq!(
+            crate::process::birth_authority(decision),
+            LeaseAuthority::Unarmed,
+            "decision {decision:?} cannot lift, so it must not clamp"
+        );
+    }
+}
